+++ Engine/Vec2.h
@@ -7,8 +7,8 @@ public:
 	Vec2( int x_in,int y_in );
 	Vec2 operator+( const Vec2& rhs ) const;
 	Vec2& operator+=( const Vec2& rhs );
-	Vec2 operator*( float rhs ) const;
-	Vec2& operator*=( float rhs );
+	Vec2 operator*( int rhs ) const;
+	Vec2& operator*=( int rhs );
 	Vec2 operator-( const Vec2& rhs ) const;
 	Vec2& operator-=( const Vec2& rhs );
 public:
